requirements.txt
aiohttp>=3.8.1,<3.9
discord-typings>=0.5.1,<1.1
pynacl>=1.5.0,<1.6.0
